Add new fn and refactor executor context
