+++ agent-node/src/executor/Executor.ts
@@ -3,6 +3,10 @@ import { ManagerInterface } from '../service/ManagerInterfaceService'
 import { FunctionGroup, getHandlers } from './AgentFunctions'
 import { Builtins, FunctionContext, Key, Wallet } from './BaseFunction'
 import { TxListener } from './TxListener'
+import { generateProposalMetadataContent } from '../utils/metadataContent/proposalMetadataContent'
+import { generateRegisterDrepMetadataContent } from '../utils/metadataContent/drepMetadataContent'
+import { generateVoteMetadataContent } from '../utils/metadataContent/voteMetadataContent'
+import { rewardAddressBech32 } from '../utils/cardano'
 
 export interface CallLog {
     function: string
@@ -37,8 +41,24 @@ export class Executor {
             },
             builtins: this.getBuiltins(wallet),
             agentName: '',
+            helpers: this.getHelpers(wallet, ''),
         }
     }
+
+    getHelpers(wallet: any, agentName: string) {
+        const context = this.functionContext
+        return {
+            generateProposalMetadataContent: () =>
+                generateProposalMetadataContent(agentName),
+            generateDrepMetadataContent: () =>
+                generateRegisterDrepMetadataContent(
+                    agentName,
+                    wallet.paymentKey.private
+                ),
+            generateVoteMetadataContent: () => generateVoteMetadataContent(),
+        }
+    }
+
     makeWallet(walletDetails: AgentWalletDetails): Wallet {
         const txSubmissionHold: any[] = []
         let isProcessing: boolean = false
@@ -64,15 +84,21 @@ export class Executor {
                 throw new Error('Key.signRaw is not implemented')
             },
         }
+        const rewardAddress = rewardAddressBech32(
+            0,
+            walletDetails.stake_verification_key_hash
+        )
         console.log(
-            'Account keys received : address =>',
-            walletDetails.agent_address
+            'Account keys received : addresses =>',
+            walletDetails.agent_address,
+            rewardAddress
         )
         return {
             address: walletDetails.agent_address,
             paymentKey: paymentKey,
             stakeKey: stakeKey,
             drepId: walletDetails.drep_id,
+            rewardAddress: rewardAddress,
             buildAndSubmit: (spec: any, stakeSigning?: boolean) => {
                 spec.selections = [
                     walletDetails.agent_address,
@@ -147,6 +173,10 @@ export class Executor {
             this.functionContext.wallet
         )
         this.functionContext.agentName = agentName
+        this.functionContext.helpers = this.getHelpers(
+            this.functionContext.wallet,
+            agentName
+        )
         managerInterface
             .getFaucetBalance(this.functionContext.wallet.address)
             .then((balance) => {
@@ -211,6 +241,7 @@ export class Executor {
             callLog: callLog,
         }
     }
+
     getBuiltins(wallet: Wallet): Builtins {
         const builtins = this.functions.builtins
         const context = this.functionContext
@@ -233,13 +264,10 @@ export class Executor {
                 )
             },
             loadFunds: async (amount: number): Promise<any> => {
-                return await this.rpcInterface.loadFunds(
-                    context.wallet.address,
-                    amount
-                )
+                return await this.rpcInterface.loadFunds(wallet.address, amount)
             },
-            saveMetadata: async (fileName, content) => {
-                return await this.rpcInterface.saveMetadata(fileName, content)
+            saveMetadata: async (content) => {
+                return await this.rpcInterface.saveMetadata(content)
             },
             ...updatedBuiltins,
         } as Builtins
