+++ agent-node/src/functions/proposalNewConstitution.ts
@@ -1,6 +1,4 @@
 import { FunctionContext } from '../executor/BaseFunction'
-import { rewardAddressBech32 } from '../utils/cardano'
-import * as blake from 'blakejs'
 
 export default async function handler(
     context: FunctionContext,
@@ -8,19 +6,21 @@ export default async function handler(
     newConstitution: any,
     guardRailScript: any
 ) {
-    const metadata = JSON.stringify(generateProposalMetadata(context.agentName))
-    const hash = blake.blake2bHex(metadata, undefined, 32)
-    const url = await context.builtins.saveMetadata(hash, metadata)
-    const rewardAddress = rewardAddressBech32(
-        0,
-        context.wallet.stakeKey.pubKeyHash
+    const { dataHash, url } = await context.builtins.saveMetadata(
+        context.helpers.generateProposalMetadataContent()
     )
+    const anchorData =
+        anchor['url'] && anchor['dataHash'] ? anchor : { url, dataHash }
+    const newConstitutionData =
+        newConstitution['url'] && newConstitution['dataHash']
+            ? newConstitution
+            : { url, dataHash }
     const req = {
         proposals: [
             {
-                anchor: anchor || { url, dataHash: hash },
-                newConstitution: newConstitution || { url, dataHash: hash },
-                refundAccount: rewardAddress,
+                anchor: anchorData,
+                newConstitution: newConstitutionData,
+                refundAccount: context.wallet.rewardAddress,
                 guardrailscript: guardRailScript || undefined,
             },
         ],
@@ -31,71 +31,4 @@ export default async function handler(
         .catch((e) => {
             throw e
         })
-}
-
-function generateProposalMetadata(agentName: string) {
-    return {
-        '@context': {
-            CIP100: 'https://github.com/cardano-foundation/CIPs/blob/master/CIP-0100/README.md#',
-            CIP108: 'https://github.com/cardano-foundation/CIPs/blob/master/CIP-0108/README.md#',
-            hashAlgorithm: 'CIP100:hashAlgorithm',
-            body: {
-                '@id': 'CIP108:body',
-                '@context': {
-                    references: {
-                        '@id': 'CIP108:references',
-                        '@container': '@set',
-                        '@context': {
-                            GovernanceMetadata:
-                                'CIP100:GovernanceMetadataReference',
-                            Other: 'CIP100:OtherReference',
-                            label: 'CIP100:reference-label',
-                            uri: 'CIP100:reference-uri',
-                            referenceHash: {
-                                '@id': 'CIP108:referenceHash',
-                                '@context': {
-                                    hashDigest: 'CIP108:hashDigest',
-                                    hashAlgorithm: 'CIP100:hashAlgorithm',
-                                },
-                            },
-                        },
-                    },
-                    title: 'CIP108:title',
-                    abstract: 'CIP108:abstract',
-                    motivation: 'CIP108:motivation',
-                    rationale: 'CIP108:rationale',
-                },
-            },
-            authors: {
-                '@id': 'CIP100:authors',
-                '@container': '@set',
-                '@context': {
-                    name: 'http://xmlns.com/foaf/0.1/name',
-                    witness: {
-                        '@id': 'CIP100:witness',
-                        '@context': {
-                            witnessAlgorithm: 'CIP100:witnessAlgorithm',
-                            publicKey: 'CIP100:publicKey',
-                            signature: 'CIP100:signature',
-                        },
-                    },
-                },
-            },
-        },
-        authors: [],
-        hashAlgorithm: 'blake2b-256',
-        body: {
-            abstract: `This proposal is created automatically by agent: ${agentName}`,
-            motivation:
-                'This proposal is automatically generated by autonomous-agent-testing.',
-            references: [
-                {
-                    '@type': 'Other',
-                    label: 'autonomous-agent-testing',
-                    uri: 'https://cardanoapi.github.io/autonomous-agents/archietecture_docusaurus/docs/architecture',
-                },
-            ],
-            title: 'Autonomous-agent-testing Proposal Metadata',
-        },
-    }
 }
