+++ agent-node/src/functions/updateCommittee.ts
@@ -2,7 +2,10 @@ import { FunctionContext } from '../executor/BaseFunction'
 
 export default async function handler(
     context: FunctionContext,
-    anchor: Record<string, any>
+    anchor: Record<string, any>,
+    quorum: Record<string, number>,
+    add: Record<any, any>,
+    remove: Array<any>
 ) {
     const { dataHash, url } = await context.builtins.saveMetadata(
         context.helpers.generateProposalMetadataContent()
@@ -14,10 +17,13 @@ export default async function handler(
             {
                 refundAccount: context.wallet.rewardAddress,
                 anchor: anchorData,
+                updatecommittee: {
+                    add: add,
+                    remove: remove,
+                    qourum: quorum,
+                },
             },
         ],
     }
-    return await context.wallet.buildAndSubmit(req).catch((e) => {
-        throw e
-    })
+    return await context.wallet.buildAndSubmit(req, false)
 }
